Fix Combination

+++ src/Combination/Combination.hpp
@@ -9,7 +9,7 @@
 
 class Combination : public Score {
     protected :
-        float score;
+        double score;
         vector<Card> totalHand;
         // int highestCombo; // 0 : high card, 1 : pair, 2 : two pair, ..., 8 : straight flush
